# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para interagir com o Redirectify",
        "title": "API do Redirectify",
        "contact": {
            "name": "Equipe da DevOps (Pablo Eduardo, Guilherme Bernardo e Eduardo Henrique)",
            "email": "comercialdevops@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/links/historico": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Histórico"
                ],
                "summary": "Retorna o histórico de ações relativas a links no sistema",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/HistoricoLink"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            }
        },
...
        "/planos_de_assinatura/historico": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Histórico"
                ],
                "summary": "Retorna o histórico de ações relativas a planos de assinatura no sistema",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/HistoricoPlanoDeAssinatura"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            }
        },
...
        "/redirecionadores/historico": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Histórico"
                ],
                "summary": "Retorna o histórico de ações relativas a redirecionadores no sistema",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/HistoricoRedirecionador"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            }
        },
...
        "/usuarios/historico": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Histórico"
                ],
                "summary": "Retorna o histórico de ações relativas a usuários no sistema",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/HistoricoUsuario"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            }
        },
...
        "/usuarios_temporarios/historico": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Histórico"
                ],
                "summary": "Retorna o histórico de ações relativas a usuários temporários no sistema",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/HistoricoUsuarioTemporario"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Erro": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
        "HistoricoLink": {
            "type": "object",
            "properties": {
                "atualizado_em": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "bitmask": {
                    "type": "integer"
                },
                "criado_em": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "criado_em_": {
                    "type": "string"
                },
                "id": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "link": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "plataforma": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "redirecionador": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "removido_em": {
                    "type": "string"
                },
                "rowid": {
                    "type": "integer"
                },
                "versao": {
                    "type": "integer"
                }
            }
        },
        "HistoricoPlanoDeAssinatura": {
            "type": "object",
            "properties": {
                "atualizado_em": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "bitmask": {
                    "type": "integer"
                },
                "criado_em": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "criado_em_": {
                    "type": "string"
                },
                "id": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "limite_links_mensal": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "nome": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "ordenacao_aleatoria_links": {
                    "$ref": "#/definitions/sql.NullBool"
                },
                "removido_em": {
                    "type": "integer"
                },
                "rowid": {
                    "type": "integer"
                },
                "valor_mensal": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "versao": {
                    "type": "integer"
                }
            }
        },
        "HistoricoRedirecionador": {
            "type": "object",
            "properties": {
                "atualizado_em": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "bitmask": {
                    "type": "integer"
                },
                "codigo_hash": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "criado_em": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "criado_em_": {
                    "type": "string"
                },
                "id": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "nome": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "ordem_de_redirecionamento": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "removido_em": {
                    "type": "string"
                },
                "rowid": {
                    "type": "integer"
                },
                "usuario": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "versao": {
                    "type": "integer"
                }
            }
        },
        "HistoricoUsuario": {
            "type": "object",
            "properties": {
                "atualizado_em": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "bitmask": {
                    "type": "integer"
                },
                "cpf": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "criado_em": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "criado_em_": {
                    "type": "string"
                },
                "data_de_nascimento": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "email": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "id": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "nome": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "nome_de_usuario": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "plano_de_assinatura": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "removido_em": {
                    "type": "string"
                },
                "rowid": {
                    "type": "integer"
                },
                "senha": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "versao": {
                    "type": "integer"
                }
            }
        },
        "HistoricoUsuarioTemporario": {
            "type": "object",
            "properties": {
                "bitmask": {
                    "type": "integer"
                },
                "cpf": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "criado_em": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "criado_em_": {
                    "type": "string"
                },
                "email": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "id": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "nome": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "nome_de_usuario": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "plano_de_assinatura": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "removido_em": {
                    "type": "string"
                },
                "rowid": {
                    "type": "integer"
                },
                "versao": {
                    "type": "integer"
                }
            }
        },
...
        "sql.NullBool": {
            "type": "object",
            "properties": {
                "bool": {
                    "type": "boolean"
                },
                "valid": {
                    "description": "Valid is true if Bool is not NULL",
                    "type": "boolean"
                }
            }
        },
        "sql.NullInt64": {
            "type": "object",
            "properties": {
                "int64": {
                    "type": "integer"
                },
                "valid": {
                    "description": "Valid is true if Int64 is not NULL",
                    "type": "boolean"
                }
            }
        },
        "sql.NullString": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}